requirements.txt
absl-py==2.3.1
accelerate==1.12.0
adagio==0.2.6
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-cors==0.8.1
aiosignal==1.4.0
alembic==1.17.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.1
attrs==25.4.0
autogluon==1.5.0
autogluon.common==1.5.0
autogluon.core==1.5.0
autogluon.features==1.5.0
autogluon.multimodal==1.5.0
autogluon.tabular==1.5.0
autogluon.timeseries==1.5.0
beartype==0.22.9
beautifulsoup4==4.14.3
blis==1.3.3
boto3==1.42.17
botocore==1.42.17
cachetools==6.2.4
catalogue==2.0.10
catboost==1.2.8
certifi==2025.11.12
charset-normalizer==3.4.4
chronos-forecasting==2.2.2
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
cloudpathlib==0.23.0
cloudpickle==3.1.2
cmake==4.2.0
colorama==0.4.6
colorful==0.5.8
colorlog==6.10.1
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
coreforecast==0.0.16
cycler==0.12.1
cymem==2.0.13
dacite==1.9.2
datasets==4.0.0
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.4.0
einops==0.8.1
einx==0.3.0
evaluate==0.4.6
executing==2.2.1
fastai==2.8.6
fastcore==1.9.8
fastdownload==0.0.7
fastprogress==1.0.5
fasttransform==0.0.2
filelock==3.20.0
filetype==1.2.0
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.3.0
fugue==0.9.3
future==1.0.0
gdown==5.2.0
gluonts==0.16.2
google-api-core==2.28.1
google-auth==2.45.0
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.3.0
grpcio==1.76.0
hatchling==1.28.0
hf-xet==1.2.0
huggingface-hub==0.36.0
hyperopt==0.2.7
idna==3.11
ImageHash==4.3.2
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonschema==4.23.0
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
lightgbm @ file:///home/ramcharan/ml/construction/LightGBM/python-package
lightning==2.5.6
lightning-utilities==0.15.2
llvmlite==0.45.1
loguru==0.7.3
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.2.1
mdurl==0.1.2
minify_html==0.18.1
mlforecast==0.14.0
model-index==0.1.11
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multimethod==1.12
multiprocess==0.70.16
murmurhash==1.0.15
narwhals==2.14.0
nest-asyncio==1.6.0
networkx==3.6.1
ninja==1.13.0
nlpaug==1.1.11
nltk==3.9.2
numba==0.62.1
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencensus==0.11.4
opencensus-context==0.1.3
opendatalab==0.0.10
openmim==0.3.9
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
openxlab==0.0.11
optuna==4.6.0
ordered-set==4.1.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
parameterized==0.9.0
parso==0.8.5
pathspec==0.12.1
patsy==1.0.2
pbr==7.0.3
pdf2image==1.17.0
peft==0.17.1
pexpect==4.9.0
phik==0.12.5
pillow==11.3.0
platformdirs==4.5.1
plotly==6.5.0
pluggy==1.6.0
plum-dispatch==2.6.1
preshed==3.0.12
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
py-spy==0.4.1
py4j==0.10.9.9
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.0
pytorch-metric-learning==2.8.1
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.3
pyzmq==27.1.0
rasterio==1.4.4
ray==2.52.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scikit_build_core==0.11.6
scipy==1.16.3
seaborn==0.13.2
sentencepiece==0.2.1
seqeval==1.2.2
setuptools==70.2.0
six==1.17.0
smart_open==7.5.0
soupsieve==2.8.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.45
srsly==2.5.2
stack-data==0.6.3
statsforecast==2.0.1
statsmodels==0.14.6
stevedore==5.4.1
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
text-unidecode==1.3
thinc==8.3.10
threadpoolctl==3.6.0
tifffile==2025.12.12
timm==1.0.3
tokenizers==0.22.1
toolz==0.12.1
torch==2.6.0+cu124
torch-geometric==2.7.0
torchaudio==2.6.0+cu124
torchmetrics==1.7.4
torchvision==0.21.0+cu124
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
triad==1.0.0
triton==3.2.0
trove-classifiers==2025.12.1.14
typeguard==4.4.4
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
utilsforecast==0.2.11
virtualenv==20.35.4
visions==0.8.1
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
Werkzeug==3.1.4
window_ops==0.0.15
wordcloud==1.9.5
wrapt==2.0.1
xgboost==3.1.2
xxhash==3.6.0
yarl==1.22.0
ydata-profiling==4.18.0
zipp==3.23.0
